requirements.txt
streamlit
langchain
groq
langchain-groq
python-dotenv